+++ src/aks-preview/azext_aks_preview/tests/latest/test_managed_cluster_decorator.py
@@ -944,7 +944,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
         )
         self.assertEqual(ctx_3.get_enable_cilium_dataplane(), False)
 
-    def test_mc_get_enable_advanced_network_observability(self):
+    def test_mc_get_acns(self):
         # Default, not set.
         ctx_1 = AKSPreviewManagedClusterContext(
             self.cmd,
@@ -952,307 +952,125 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        self.assertEqual(ctx_1.get_enable_advanced_network_observability(), None)
+        self.assertEqual(ctx_1.get_acns(), None)
 
         # Flag set to True.
         ctx_2 = AKSPreviewManagedClusterContext(
             self.cmd,
             AKSManagedClusterParamDict(
                 {
-                    "enable_advanced_network_observability": True,
+                    "enable_acns": True,
                 }
             ),
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        self.assertEqual(ctx_2.get_enable_advanced_network_observability(), True)
+        self.assertEqual(ctx_2.get_acns(), True)
 
         # Flag set to True.
         ctx_3 = AKSPreviewManagedClusterContext(
             self.cmd,
             AKSManagedClusterParamDict(
                 {
-                    "enable_advanced_network_observability": True,
+                    "enable_acns": True,
                 }
             ),
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(ctx_3.get_enable_advanced_network_observability(), True)
+        self.assertEqual(ctx_3.get_acns(), True)
 
         # Flag set to True and False.
         ctx_4 = AKSPreviewManagedClusterContext(
             self.cmd,
             AKSManagedClusterParamDict(
                 {
-                    "enable_advanced_network_observability": True,
-                    "disable_advanced_network_observability": True,
+                    "enable_acns": True,
+                    "disable_acns": True,
                 }
             ),
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        # fail on get_enable_network_observability mutual exclusive error
+        # fail on get_acns mutual exclusive error
         with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_4.get_enable_advanced_network_observability()
+            ctx_4.get_acns()
 
         # Flag set to False.
         ctx_5 = AKSPreviewManagedClusterContext(
             self.cmd,
             AKSManagedClusterParamDict(
                 {
-                    "disable_advanced_network_observability": True,
+                    "disable_acns": True,
                 }
             ),
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(ctx_5.get_enable_advanced_network_observability(), False)
+        self.assertEqual(ctx_5.get_acns(), False)
 
-    def test_mc_get_advanced_networking_observability_tls_management(self):
-        # Default, not set.
-        ctx_1 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict({}),
-            self.models,
-            decorator_mode=DecoratorMode.CREATE,
-        )
-        self.assertEqual(ctx_1.get_advanced_networking_observability_tls_management(), None)
-
-        # Flag set to Managed.
-        ctx_2 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "enable_advanced_network_observability": True,
-                    "advanced_networking_observability_tls_management": "Managed",
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.CREATE,
-        )
-        self.assertEqual(ctx_2.get_advanced_networking_observability_tls_management(), "Managed")
-
-        # Flag set to None.
-        ctx_3 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "enable_advanced_network_observability": True,
-                    "advanced_networking_observability_tls_management": "None",
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.UPDATE,
-        )
-        self.assertEqual(ctx_3.get_advanced_networking_observability_tls_management(), "None")
-
-        # Flag set to Managed when Observability is disabled.
-        ctx_4 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "disable_advanced_network_observability": True,
-                    "advanced_networking_observability_tls_management": "Managed",
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.UPDATE,
-        )
-        with self.assertRaises(ArgumentUsageError):
-            ctx_4.get_advanced_networking_observability_tls_management()
-
-        # Flag set to None when Observability is disabled.
-        ctx_4 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "disable_advanced_network_observability": True,
-                    "advanced_networking_observability_tls_management": "None",
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.UPDATE,
-        )
-        with self.assertRaises(ArgumentUsageError):
-            ctx_4.get_advanced_networking_observability_tls_management()
-
-    def test_mc_get_enable_fqdn_policy(self):
-        # Default, not set.
-        ctx_1 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict({}),
-            self.models,
-            decorator_mode=DecoratorMode.CREATE,
-        )
-        self.assertEqual(ctx_1.get_enable_fqdn_policy(), None)
-
-        # Flag set to True.
-        ctx_2 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "enable_fqdn_policy": True,
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.CREATE,
-        )
-        self.assertEqual(ctx_2.get_enable_fqdn_policy(), True)
-
-        # Flag set to True.
-        ctx_3 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "enable_fqdn_policy": True,
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.UPDATE,
-        )
-        self.assertEqual(ctx_3.get_enable_fqdn_policy(), True)
-
-        # Flag set to True and False.
-        ctx_4 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "enable_fqdn_policy": True,
-                    "disable_fqdn_policy": True,
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.UPDATE,
-        )
-        # fail on get_enable_fqdn_policy mutual exclusive error
-        with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_4.get_enable_fqdn_policy()
-
-        # Flag set to False.
-        ctx_5 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "disable_fqdn_policy": True,
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.UPDATE,
-        )
-        self.assertEqual(ctx_5.get_enable_fqdn_policy(), False)
-
-        # Check behavior of advanced networking observability and security.
-        # Enable obs, enable fqdn.
+        # Illegal flags enable acns and disable acns
         ctx_6 = AKSPreviewManagedClusterContext(
             self.cmd,
             AKSManagedClusterParamDict(
                 {
-                    "enable_advanced_network_observability": True,
-                    "enable_fqdn_policy": True,
+                    "enable_acns": True,
+                    "disable_acns": True,
                 }
             ),
             self.models,
-            decorator_mode=DecoratorMode.CREATE,
+            decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(ctx_6.get_enable_fqdn_policy(), True)
-        self.assertEqual(ctx_6.get_enable_advanced_network_observability(), True)
+        # fail on get_acns mutual exclusive error
+        with self.assertRaises(MutuallyExclusiveArgumentError):
+            ctx_6.get_acns()
 
+        # Illegal flags enable acns, disable acns security, disable acns observability
         ctx_7 = AKSPreviewManagedClusterContext(
             self.cmd,
             AKSManagedClusterParamDict(
                 {
-                    "enable_advanced_network_observability": True,
-                    "enable_fqdn_policy": True,
+                    "enable_acns": True,
+                    "disable_acns_observability": True,
+                    "disable_acns_security": True,
                 }
             ),
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(ctx_7.get_enable_fqdn_policy(), True)
-        self.assertEqual(ctx_7.get_enable_advanced_network_observability(), True)
-
-    def test_mc_get_enable_acns(self):
-        # Default, not set.
-        ctx_1 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict({}),
-            self.models,
-            decorator_mode=DecoratorMode.CREATE,
-        )
-        self.assertEqual(ctx_1.get_enable_acns(), None)
-
-        # Flag set to True.
-        ctx_2 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "enable_acns": True,
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.CREATE,
-        )
-        self.assertEqual(ctx_2.get_enable_acns(), True)
-
-        # Flag set to True.
-        ctx_3 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "enable_acns": True,
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.UPDATE,
-        )
-        self.assertEqual(ctx_3.get_enable_acns(), True)
-
-        # Flag set to True and False.
-        ctx_4 = AKSPreviewManagedClusterContext(
-            self.cmd,
-            AKSManagedClusterParamDict(
-                {
-                    "enable_acns": True,
-                    "disable_acns": True,
-                }
-            ),
-            self.models,
-            decorator_mode=DecoratorMode.UPDATE,
-        )
-        # fail on get_enable_acns mutual exclusive error
+        # fail on get_acns mutual exclusive error
         with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_4.get_enable_acns()
+            ctx_7.get_acns()
 
-        # Flag set to False.
-        ctx_5 = AKSPreviewManagedClusterContext(
+        # Illegal flags disable acns observability without specifying enable acns
+        ctx_8 = AKSPreviewManagedClusterContext(
             self.cmd,
             AKSManagedClusterParamDict(
                 {
-                    "disable_acns": True,
+                    "disable_acns_observability": True,
                 }
             ),
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(ctx_5.get_enable_acns(), False)
+        # fail on get_acns mutual exclusive error
+        with self.assertRaises(RequiredArgumentMissingError):
+            ctx_8.get_acns()
 
-        # Illegal flags
-        ctx_6 = AKSPreviewManagedClusterContext(
+        # Illegal flags disable acns security without specifying enable acns
+        ctx_9 = AKSPreviewManagedClusterContext(
             self.cmd,
             AKSManagedClusterParamDict(
                 {
-                    "enable_acns": True,
-                    "disable_fqdn_policy": True,
-                    "enable_advanced_network_observability": True,
+                    "disable_acns_security": True,
                 }
             ),
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        # fail on get_enable_acns mutual exclusive error
-        with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_6.get_enable_acns()
+        # fail on get_acns mutual exclusive error
+        with self.assertRaises(RequiredArgumentMissingError):
+            ctx_9.get_acns()
 
     def test_get_enable_managed_identity(self):
         # custom value
@@ -5587,26 +5405,6 @@ class AKSPreviewManagedClusterCreateDecoratorTestCase(unittest.TestCase):
         )
         self.assertEqual(dec_mc_2, expect_mc_2)
 
-        dec_3 = AKSPreviewManagedClusterCreateDecorator(
-            self.cmd,
-            self.client,
-            {"uptime_sla": True},
-            CUSTOM_MGMT_AKS_PREVIEW,
-        )
-        mc_3 = self.models.ManagedCluster(
-            location="test_location",
-            sku=None,
-        )
-        dec_3.context.attach_mc(mc_3)
-        dec_mc_3 = dec_3.set_up_sku(mc_3)
-        baseSKU = self.models.ManagedClusterSKU(name="Base", tier="Standard")
-        expect_mc_3 = self.models.ManagedCluster(
-            location="test_location",
-            sku=baseSKU,
-            kind="Base",
-        )
-        self.assertEqual(dec_mc_3, expect_mc_3)
-
 
 class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
     def setUp(self):
@@ -8644,83 +8442,6 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
             kind="Base",
         )
         self.assertEqual(dec_mc_2, expect_mc_2)
-
-        dec_3 =  AKSPreviewManagedClusterUpdateDecorator(
-            self.cmd,
-            self.client,
-            {
-                "uptime_sla": True,
-                "no_uptime_sla": True,
-            },
-            CUSTOM_MGMT_AKS_PREVIEW,
-        )
-        mc_3 = self.models.ManagedCluster(
-            location="test_location",
-            sku=self.models.ManagedClusterSKU(
-                name="Base",
-                tier="Free",
-            ),
-        )
-        dec_3.context.attach_mc(mc_3)
-        # fail on mutually exclusive uptime_sla and no_uptime_sla
-        with self.assertRaises(MutuallyExclusiveArgumentError):
-            dec_3.update_sku(mc_3)
-
-        dec_4 = AKSPreviewManagedClusterUpdateDecorator(
-            self.cmd,
-            self.client,
-            {
-                "uptime_sla": False,
-                "no_uptime_sla": True,
-            },
-            CUSTOM_MGMT_AKS_PREVIEW,
-        )
-        mc_4 = self.models.ManagedCluster(
-            location="test_location",
-            sku=self.models.ManagedClusterSKU(
-                name="Base",
-                tier="Standard",
-            ),
-        )
-        dec_4.context.attach_mc(mc_4)
-        dec_mc_4 = dec_4.update_sku(mc_4)
-        ground_truth_mc_4 = self.models.ManagedCluster(
-            location="test_location",
-            sku=self.models.ManagedClusterSKU(
-                name="Base",
-                tier="Free",
-            ),
-            kind="Base",
-        )
-        self.assertEqual(dec_mc_4, ground_truth_mc_4)
-
-        dec_5 = AKSPreviewManagedClusterUpdateDecorator(
-            self.cmd,
-            self.client,
-            {
-                "uptime_sla": True,
-                "no_uptime_sla": False,
-            },
-            CUSTOM_MGMT_AKS_PREVIEW,
-        )
-        mc_5 = self.models.ManagedCluster(
-            location="test_location",
-            sku=self.models.ManagedClusterSKU(
-                name="Base",
-                tier="Free",
-            ),
-        )
-        dec_5.context.attach_mc(mc_5)
-        dec_mc_5 = dec_5.update_sku(mc_5)
-        ground_truth_mc_5 = self.models.ManagedCluster(
-            location="test_location",
-            sku=self.models.ManagedClusterSKU(
-                name="Base",
-                tier="Standard",
-            ),
-            kind="Base",
-        )
-        self.assertEqual(dec_mc_5, ground_truth_mc_5)
 
     def test_setup_supportPlan(self):
         # default value in `aks_create`