+++ src/connectedk8s/azext_connectedk8s/custom.py
@@ -294,6 +294,10 @@ def create_connectedk8s(
     ):
         lowbandwidth = True
 
+    azure_local_disconnected = False
+    if os.getenv("AZURE_LOCAL_DISCONNECTED") == "true":
+        azure_local_disconnected = True
+
     # Install kubectl and helm
     try:
         kubectl_client_location = install_kubectl_client()
@@ -309,8 +313,8 @@ def create_connectedk8s(
     # Pre onboarding checks
     diagnostic_checks = "Failed"
     try:
-        # if aks_hci lowbandwidth scenario skip, otherwise continue to perform pre-onboarding check
-        if not lowbandwidth:
+        # if aks_hci lowbandwidth scenario or Azure local disconnected, skip, otherwise continue pre-onboarding check.
+        if not azure_local_disconnected and not lowbandwidth:
             print(f"Step: {utils.get_utctimestring()}: Starting Pre-onboarding-check")
             batchv1_api_instance = kube_client.BatchV1Api()
             storage_space_available = True
@@ -389,7 +393,7 @@ def create_connectedk8s(
         raise ManualInterrupt("Process terminated externally.")
 
     # If the checks didnt pass then stop the onboarding
-    if diagnostic_checks != consts.Diagnostic_Check_Passed and lowbandwidth is False:
+    if diagnostic_checks != consts.Diagnostic_Check_Passed and not azure_local_disconnected and not lowbandwidth:
         if storage_space_available:
             logger.warning(
                 "The pre-check result logs logs have been saved at this path: "
@@ -422,7 +426,7 @@ def create_connectedk8s(
         )
         raise ValidationError(err_msg)
 
-    if lowbandwidth is False:
+    if not azure_local_disconnected and not lowbandwidth:
         print(
             f"Step: {utils.get_utctimestring()}: The required pre-checks for onboarding have succeeded."
         )
@@ -899,6 +903,7 @@ def create_connectedk8s(
     )
 
     helm_content_values = helm_values_dp["helmValuesContent"]
+    aad_identity_principal_id = put_cc_response.identity.principal_id,
 
     # Substitute any protected helm values as the value for that will be 'redacted-<feature>-<protectedSetting>'
     for helm_parameter, helm_value in helm_content_values.items():
@@ -930,6 +935,8 @@ def create_connectedk8s(
         enable_private_link,
         arm_metadata,
         helm_content_values,
+        registry_path,
+        aad_identity_principal_id,
         onboarding_timeout,
     )
 
@@ -2267,9 +2274,11 @@ def update_connected_cluster(
     for helm_parameter, helm_value in helm_content_values.items():
         if "redacted" in helm_value:
             _, feature, protectedSetting = helm_value.split(":")
-            helm_content_values[helm_parameter] = configuration_protected_settings[
-                feature
-            ][protectedSetting]
+            helm_content_values[helm_parameter] = configuration_protected_settings[feature][protectedSetting]
+
+    # Disable proxy if disable_proxy flag is set
+    if disable_proxy:
+        helm_content_values["global.isProxyEnabled"] = "False"
 
     # Disable proxy if disable_proxy flag is set
     if disable_proxy:
@@ -3355,10 +3364,25 @@ def merge_kubernetes_configurations(
         except (KeyError, TypeError):
             continue
 
-    handle_merge(existing, addition, "clusters", replace)
-    handle_merge(existing, addition, "users", replace)
-    handle_merge(existing, addition, "contexts", replace)
-    existing["current-context"] = addition["current-context"]
+    if addition is None:
+        telemetry.set_exception(
+            exception="Failed to load additional configuration",
+            fault_type=consts.Failed_To_Load_K8s_Configuration_Fault_Type,
+            summary="failed to load additional configuration from {}".format(
+                addition_file
+            ),
+        )
+        raise CLIInternalError(
+            f"Failed to load additional configuration from {addition_file}"
+        )
+
+    if existing is None:
+        existing = addition
+    else:
+        handle_merge(existing, addition, "clusters", replace)
+        handle_merge(existing, addition, "users", replace)
+        handle_merge(existing, addition, "contexts", replace)
+        existing["current-context"] = addition["current-context"]
 
     # check that ~/.kube/config is only read- and writable by its owner
     if platform.system() != "Windows":
@@ -3605,6 +3629,13 @@ def client_side_proxy_wrapper(
     # initializations
     user_type = "sat"
     creds = ""
+    dict_file = {
+        "server": {
+            "httpPort": int(client_proxy_port),
+            "httpsPort": int(api_server_port)
+        },
+        "identity": {"tenantID": tenant_id}
+    }
 
     # if service account token is not passed
     if token is None:
@@ -3615,35 +3646,9 @@ def client_side_proxy_wrapper(
 
         dict_file: dict[str, Any]
         if user_type == "user":
-            dict_file = {
-                "server": {
-                    "httpPort": int(client_proxy_port),
-                    "httpsPort": int(api_server_port),
-                },
-                "identity": {
-                    "tenantID": tenant_id,
-                    "clientID": consts.CLIENTPROXY_CLIENT_ID,
-                },
-            }
+            dict_file["identity"]["clientID"] = consts.CLIENTPROXY_CLIENT_ID
         else:
-            dict_file = {
-                "server": {
-                    "httpPort": int(client_proxy_port),
-                    "httpsPort": int(api_server_port),
-                },
-                "identity": {
-                    "tenantID": tenant_id,
-                    "clientID": account["user"]["name"],
-                },
-            }
-
-        if cloud == "DOGFOOD":
-            dict_file["cloud"] = "AzureDogFood"
-
-        if cloud == consts.Azure_ChinaCloudName:
-            dict_file["cloud"] = "AzureChinaCloud"
-        elif cloud == consts.Azure_USGovCloudName:
-            dict_file["cloud"] = "AzureUSGovernmentCloud"
+            dict_file["identity"]["clientID"] = account["user"]["name"]
 
         if not utils.is_cli_using_msal_auth():
             # Fetching creds
@@ -3685,17 +3690,27 @@ def client_side_proxy_wrapper(
 
             if user_type != "user":
                 dict_file["identity"]["clientSecret"] = creds
+
+    if cloud == "DOGFOOD":
+        dict_file["cloud"] = "AzureDogFood"
+    elif cloud == consts.Azure_ChinaCloudName:
+        dict_file["cloud"] = "AzureChinaCloud"
+    elif cloud == consts.Azure_USGovCloudName:
+        dict_file["cloud"] = "AzureUSGovernmentCloud"
     else:
-        dict_file = {
-            "server": {
-                "httpPort": int(client_proxy_port),
-                "httpsPort": int(api_server_port),
-            }
-        }
-        if cloud == consts.Azure_ChinaCloudName:
-            dict_file["cloud"] = "AzureChinaCloud"
-        elif cloud == consts.Azure_USGovCloudName:
-            dict_file["cloud"] = "AzureUSGovernmentCloud"
+        dict_file["cloud"] = cloud
+
+    # Azure local configurations.
+    arm_metadata = utils.get_metadata(cmd.cli_ctx.cloud.endpoints.resource_manager)
+    if "dataplaneEndpoints" in arm_metadata:
+        dict_file["cloudConfig"] = {}
+        dict_file["cloudConfig"]["resourceManagerEndpoint"] = arm_metadata["resourceManager"]
+        relay_endpoint_suffix = arm_metadata["suffixes"]["relayEndpointSuffix"]
+        if relay_endpoint_suffix[0] == ".":
+            dict_file["cloudConfig"]["serviceBusEndpointSuffix"] = (relay_endpoint_suffix)[1:]
+        else:
+            dict_file["cloudConfig"]["serviceBusEndpointSuffix"] = relay_endpoint_suffix
+        dict_file["cloudConfig"]["activeDirectoryEndpoint"] = arm_metadata["authentication"]["loginEndpoint"]
 
     telemetry.set_debug_info("User type is ", user_type)
 
@@ -4530,6 +4545,9 @@ def install_kubectl_client() -> str:
         f"Step: {utils.get_utctimestring()}: Install Kubectl client if it does not exist"
     )
     # Return kubectl client path set by user
+    if os.getenv("KUBECTL_CLIENT_PATH"):
+        return os.getenv("KUBECTL_CLIENT_PATH")
+
     try:
         # Fetching the current directory where the cli installs the kubectl executable
         home_dir = os.path.expanduser("~")