+++ src/network-manager/azext_network_manager/tests/latest/test_network_scenario.py
@@ -666,11 +666,11 @@ class NetworkScenarioTest(ScenarioTest):
                  '--scope-accesses "SecurityAdmin" '
                  '--network-manager-scopes '
                  'subscriptions={sub} '
-                 '-l eastus2 '
+                 '-l eastus2euap '
                  '--resource-group {rg}')
 
-        self.cmd('az network manager ipam-pool create --name {pool_name} --manager-name {manager_name} --resource-group {rg} --address-prefixes "["10.0.0.0/16"]" --location "eastus2euap"',
-                 self.check('name', '{pool_name}'))
+        self.cmd('az network manager ipam-pool create --name {pool_name} --manager-name {manager_name} --resource-group {rg} --address-prefixes "[\"10.0.0.0/16\"]" --location "eastus2euap"')
+        self.cmd('az resource wait --created --name {pool_name} --resource-group {rg} --resource-type "Microsoft.Network/networkManagers/ipamPools" --timeout 60')
 
         self.cmd('az network manager ipam-pool show --name {pool_name} --manager-name {manager_name} --resource-group {rg}',
                  self.check('name', '{pool_name}'))
@@ -680,8 +680,10 @@ class NetworkScenarioTest(ScenarioTest):
         self.cmd('az network manager ipam-pool show --name {pool_name} --manager-name {manager_name} --resource-group {rg}',
                  self.check('properties.description', 'updated desc'))
         
-        self.cmd('az network manager ipam-pool static-cidr create --name {staticCidr_name} --pool-name {pool_name} --manager-name {manager_name} --resource-group {rg} --number-of-ip-addresses-to-allocate {num_to_allocate}',
-                 self.check('name', '{staticCidr_name}'))
+        self.cmd('az network manager ipam-pool static-cidr create --name {staticCidr_name} --pool-name {pool_name} --manager-name {manager_name} --resource-group {rg} --number-of-ip-addresses-to-allocate {num_to_allocate}')
+        self.cmd('az resource wait --created --name {staticCidr_name} --resource-group {rg} --resource-type "Microsoft.Network/networkManagers/ipamPools/staticCidrs" --timeout 60')
+
+
         self.cmd('az network manager ipam-pool static-cidr list --pool-name {pool_name} --manager-name {manager_name} --resource-group {rg}',
                  self.check('length(@)', 1))
         self.cmd('az network manager ipam-pool static-cidr show --name {staticCidr_name} --pool-name {pool_name} --manager-name {manager_name} --resource-group {rg}',