+++ src/virtual-network-tap/HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+1.0.0b2
++++++
+* Migrate to AAZ
+* Add confirmation prompt for network vnet tap delete
+* Add confirmation prompt for network nic vtap-config delete
+
 1.0.0b1
 +++++
 * Replace msrestazure with azure.core